lib/scraper/requirements.txt
# Python dependencies for scraping and analysis
requests==2.31.0
beautifulsoup4==4.12.3
openai==1.12.0
lxml==5.1.0
supabase==2.3.4
python-dotenv==1.0.0
